rename e2e tests utils file
