+++ tests/e2e_tests/test_incentive.py
@@ -10,7 +10,7 @@ from tests.e2e_tests.utils.chain_interactions import (
     register_subnet,
     wait_epoch,
 )
-from tests.e2e_tests.utils.test_utils import (
+from tests.e2e_tests.utils.e2e_test_utils import (
     setup_wallet,
     template_path,
     templates_repo,
