+++ tests/e2e_tests/test_metagraph.py
@@ -7,7 +7,7 @@ from tests.e2e_tests.utils.chain_interactions import (
     register_neuron,
     register_subnet,
 )
-from tests.e2e_tests.utils.test_utils import (
+from tests.e2e_tests.utils.e2e_test_utils import (
     setup_wallet,
 )
